requirements.txt
Django==1.8.7
django-filter==1.0.2
djangorestframework==3.6.2
python-dateutil==2.4.2
requests==2.7.0
six==1.8.0
